protection of all apis

+++ controllers/contact/delete.contact.js
@@ -9,6 +9,10 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error(`Contact ${req.params.id} not found`);
     }
+    if (contact.user_id.toString() !== req.user.id.toString()) {
+        res.status(403);
+        throw new Error(`You are not authorized to delete this contact`);
+    }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json({
         message: `Contact ${req.params.id} has been deleted`,
